requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
boto3==1.19.8
botocore==1.22.8; python_version >= '3.6'
django-address==0.2.5
django-environ==0.8.1
django-health-check==3.16.4
django-map-widgets==0.3.2
django-phone-field==1.8.1
django-storages==1.12.3
django==3.2.9
djangorestframework-gis==0.17
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.4.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
s3transfer==0.5.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.3.0
